initial doc scraper commit
Almost finished, but needs refinement and testing with the readthedocs stuff. At
the moment it will get an OAuth key from github, and then grab data about all the directories in the organisation. It looks through each repo, and clones its wiki (if it exists), and copies over any readme files as well. The next step is to get those files to all be combined into a readthedocs webpage somehow or other.